+++ CommonplaceBookAppTests/ClozeTests.swift
@@ -66,22 +66,8 @@ final class ClozeTests: XCTestCase {
     let example = """
     * Yo ?[to be](soy) de España. ¿De dónde ?[to be](es) ustedes?
     """
-    let blocks = parsingRules.parse(example)
-    XCTAssertEqual(blocks.count, 1)
-    guard let template = ClozeTemplate.extract(from: blocks).first else {
-      XCTFail("Could not load template")
-      return
-    }
-
-    do {
-      let text = try YAMLEncoder().encode(template)
-      print(text)
-      let decoder = YAMLDecoder()
-      let decoded = try decoder.decode(ClozeTemplate.self, from: text, userInfo: [.markdownParsingRules: parsingRules])
-      XCTAssertEqual(decoded.challenges.count, template.challenges.count)
-    } catch {
-      XCTFail("Unexpected error: \(error)")
-    }
+    let decoded = ClozeTemplate(rawValue: example)
+    XCTAssertEqual(decoded?.challenges.count, 2)
   }
 
   func testClozeFormatting() {