+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- *Challenge template stability*. Challenges created from parsing markdown don't have an identity. We now have code that looks for identical-or-close templates from a prior version of a `Note` to preserve the identity of that template.
+
+### Changed
+
+- Significant refactor of `ChallengeTemplate`. They are no longer `Codable` -- instead they are `RawRepresentable`.
+
 ## [0.15.0] - 2020-01-12
 
 ### Added